lesson_10/requirements.txt
selenium==4.15.0
pytest==7.4.3
allure-pytest==2.9.45
pytest-html==4.1.1
flake8==6.1.0
webdriver-manager==4.0.1